Updated tests for the apiSlice refactror
